Add some test code

+++ detection_ctpn/text_connector/text_proposal_connector.py
@@ -48,7 +48,7 @@ class TextProposalConnector:
 
         text_lines = clip_boxes(text_lines, im_size)
 
-        text_recs = np.zeros((len(text_lines), 9), np.float)
+        text_recs = np.zeros((len(text_lines), 9), np.float32)
         index = 0
         for line in text_lines:
             xmin, ymin, xmax, ymax = line[0], line[1], line[2], line[3]
